+++ linking-platform/link-datasource/src/ui/issue-like-table/column-picker/messages.ts
@@ -1,0 +1,9 @@
+import { defineMessages } from 'react-intl-next';
+
+export const columnPickerMessages = defineMessages({
+  search: {
+    id: 'linkDataSource.column-picker.search',
+    description: 'Search bar message to look for more fields',
+    defaultMessage: 'Search for fields',
+  },
+});