+++ linking-platform/link-datasource/src/ui/issue-like-table/column-picker/types.ts
@@ -1,6 +1,7 @@
 import { DatasourceResponseSchemaProperty } from '@atlaskit/linking-types';
 
 export type ColumnPickerProps = {
+  isDatasourceLoading: boolean;
   columns: DatasourceResponseSchemaProperty[];
   selectedColumnKeys: string[];
   onSelectedColumnKeysChange: (selectedColumnKeys: string[]) => void;